Updated image.tsx
The image.tsx can now accept image dimensions (width and height) as both number or string. When number is used, the pixels will be attached to the number value. When string is used, the measurement unit must be used.
For example: 
`width: 100` - the image width will be 100px
`width: '100%'` - the image width will be 100%